chore: move MockedResponse to nodeHelper
