+++ packages/opentelemetry-exporter-collector/test/node/CollectorTraceExporter.test.ts
@@ -28,7 +28,7 @@ import * as sinon from 'sinon';
 import { CollectorTraceExporter } from '../../src/platform/node';
 import { CollectorExporterConfigBase } from '../../src/types';
 import * as collectorTypes from '../../src/types';
-import { MockedResponse } from '../helper';
+import { MockedResponse } from './nodeHelpers';
 
 import {
   ensureExportTraceServiceRequestIsSet,
